src/requirements.txt
Flask
requests
pymysql
python-dotenv
boto3